NTL-12 Shift all meta in-game data to configuration. Start rune implementation

+++ src/main/java/remonone/nftilation/game/models/PlayerModel.java
@@ -11,12 +11,12 @@ public class PlayerModel {
     @Setter
     private Player reference;
     @Setter
-    private int tokens;
+    private float tokens;
     private final Map<String, Object> parameters;
     private final PriorityQueue<IDamageHandler> damageHandlers = new PriorityQueue<>();
     private final PriorityQueue<IDamageInvoker> damageInvokers = new PriorityQueue<>();
     
-    public PlayerModel(Player reference, int tokens, Map<String, Object> params) {
+    public PlayerModel(Player reference, float tokens, Map<String, Object> params) {
         this.reference = reference;
         this.tokens = tokens;
         this.parameters = params;
